Unify Buttons / Gestures for our OnePlus devices (OP2 and OP3/T)
- Removed button swap option. This option made sense for other ROMs and even on
the stock ROM (Oxygen OS) where there was no way to configure the hardware keys
but that's not the case for us. We have the option to configure hardware keys as
we see fit. Having this just confuses the user IMO, especially in two different
menus.

The TouchscreenGestureSettings class is being mapped inside Dirty Tweaks so it
makes sense to just add in the alert slider options in here under Device extras.

Change-Id: I2830416f2d3df3126e7195edca704a58ecd1eb99